# docs/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rust-web-dashboard",
    "description": "Standalone Rust web dashboard for a DCS-gRPC server (single-binary axum + tonic)",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/console": {
      "post": {
        "tags": [
          "dcs"
        ],
        "summary": "`POST /api/console` → evaluate Lua in the mission (CustomService.Eval).",
        "operationId": "console",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsoleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Lua evaluation result"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsoleBody": {
        "type": "object",
        "properties": {
          "lua": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}